+++ back/shoppingMart/src/main/java/back/shoppingMart/purchase/dto/PurchaseProductWithImageDto.java
@@ -5,17 +5,17 @@ import lombok.Data;
 @Data
 public class PurchaseProductWithImageDto {
     private String productName;
-    private int count;
-    private double totalPrice;
+    private Integer count;
+    private Integer totalPrice;
     private String productImage;
 
-    public PurchaseProductWithImageDto(String productName, int count, double totalPrice) {
+    public PurchaseProductWithImageDto(String productName, Integer count, Integer totalPrice) {
         this.productName = productName;
         this.count = count;
         this.totalPrice = totalPrice;
     }
 
-    public PurchaseProductWithImageDto(String productName, int count, double totalPrice, String productImage) {
+    public PurchaseProductWithImageDto(String productName, Integer count, Integer totalPrice, String productImage) {
         this.productName = productName;
         this.count = count;
         this.totalPrice = totalPrice;
